Support reads of binary data
The existing code always turned on termChar, which results in problems
when binary data is read since binary data may innocently contain
termChar (defaulted to '\n'). A new method BulkRead unconditionally
disables termChar.

Added debugging information conditioned on an environment variable.
Print sent and received headers.

Allow the starting tag number to be overridden in the context. This may
or may not be a useful change.

Don't hard-code the initial buffer size. Read as much as the client
wants to read, and then some.